+++ apps/java-api/src/test/java/org/scoalaonline/api/controller/LectureMaterialControllerTest.java
@@ -3,8 +3,8 @@ package org.scoalaonline.api.controller;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.scoalaonline.api.exception.LectureMaterialInvalidDocumentException;
-import org.scoalaonline.api.exception.LectureMaterialNotFoundException;
+import org.scoalaonline.api.exception.lectureMaterial.LectureMaterialInvalidDocumentException;
+import org.scoalaonline.api.exception.lectureMaterial.LectureMaterialNotFoundException;
 import org.scoalaonline.api.model.LectureMaterial;
 import org.scoalaonline.api.service.LectureMaterialService;
 import org.springframework.beans.factory.annotation.Autowired;
